requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
autopep8==1.5.7
certifi==2022.9.24
cffi==1.15.1
channels==3.0.4
channels-redis==4.0.0
charset-normalizer==2.1.1
colorama==0.4.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
daphne==3.0.2
Deprecated==1.2.13
dj-database-url==1.0.0
dj-rest-auth==2.2.5
Django==4.1.2
django-cors-headers==3.8.0
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.1
exceptiongroup==1.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
msgpack==1.0.4
mysqlclient==2.1.1
openapi-codec==1.3.2
packaging==21.3
Pillow==9.0.1
pluggy==1.0.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pytest==7.2.0
pytest-asyncio==0.20.1
pytest-django==4.5.2
python-dotenv==0.21.0
pytz==2022.5
redis==4.3.4
requests==2.28.1
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.3
stripe==5.0.0
toml==0.10.2
tomli==2.0.1
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.5
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
wrapt==1.14.1
zope.interface==5.5.0
